int to long

+++ app/src/main/java/com/example/tcwl_manage/models/services/ApiLoginService.java
@@ -14,5 +14,5 @@ public interface ApiLoginService {
     Observable<Login> getLogin (@Query("registrationId") String registrationId,
                                 @Query("appType") int appType,
                                 @Query("password") String password,
-                                @Query("mobileNo")int mobileNo);
+                                @Query("mobileNo")long mobileNo);
 }
